Conflicts:
	code/Muse/test/edu/wm/cs/muse/MuseTest.java

+++ code/Muse/test/edu/wm/cs/muse/MuseTest.java
@@ -31,6 +31,9 @@ import org.eclipse.text.edits.MalformedTreeException;
 import org.eclipse.text.edits.TextEdit;
 import org.junit.After;
 import org.junit.Test;
+import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
+import org.junit.jupiter.api.Order;
+import org.junit.jupiter.api.TestMethodOrder;
 
 import edu.wm.cs.muse.dataleak.support.Arguments;
 import edu.wm.cs.muse.dataleak.support.FileUtility;
@@ -52,6 +55,8 @@ import  edu.wm.cs.muse.mdroid.ASTHelper;
  * @author Amit Seal Ami, Liz Weech, Yang Zhang, Scott Murphy
  *
  */
+
+@TestMethodOrder(OrderAnnotation.class)
 public class MuseTest {
 
 	File expectedOutput;
@@ -95,9 +100,8 @@ public class MuseTest {
 		pw.close();
 	}
 	
-
-	
 	@Test
+	@Order(1)
 	public void reachability_operation_on_hello_world() throws Exception {
 
 		try {
@@ -118,6 +122,7 @@ public class MuseTest {
 	}
 
 	@Test
+	@Order(2)
 	public void complex_reachability_operation_on_multi_class() {
 		try {
 
@@ -138,6 +143,7 @@ public class MuseTest {
 	}
 	
 	@Test
+	@Order(3)
 	public void taint_source_operation_on_hello_world() {
 		try {
 			prepare_test_files(OperatorType.TAINTSOURCE, 1);
@@ -157,6 +163,7 @@ public class MuseTest {
 	}
 
 	@Test
+	@Order(4)
 	public void taint_sink_operation_on_hello_world() {
 
 		try {
@@ -178,6 +185,7 @@ public class MuseTest {
 	}
 
 	@Test
+	@Order(5)
 	public void scope_source_operation_on_hello_world() {
 		try {
 
@@ -220,6 +228,7 @@ public class MuseTest {
 	}
 	
 	@Test
+	@Order(7)
 	public void scope_sink_operation_on_multi_class() {
 		try {
 			prepare_test_files(OperatorType.SCOPESINK, 1);
@@ -246,9 +255,13 @@ public class MuseTest {
 		processedOutput = output;
 		//if temp_file is created by execution of muse, break it down after use
 		File file = new File("test/temp/temp_file.java");
+		if (file.exists()) {
 		file.delete();
+		}
 		File folder = new File("test/temp");
-		folder.delete();
+		if (file.exists()) {
+			folder.delete();
+		}
 	}
